Add convertKeysTo, convertToSnakeCase
This replaces convertToCamelCase with a generic function
`convertKeysTo`. This is so it can be used similarly to the
undocumented.js function `mapKeysRecursively` in `convertToSnakeCase`

I've left existing convertToCamelCase tests in place to ensure forward
compatability. These can be removed later. The new tests do omit testing
bracket and dot keys but that looks to be lodash behaviour we _probably_
dont need to test for.

https://github.com/Automattic/wp-calypso/blob/master/client/lib/wpcom-undocumented/lib/undocumented.js#L674